+++ lib/hasil_karya/controller/hasil_karya_controller.dart
@@ -140,8 +140,12 @@ class HkController extends StateNotifier<bool> {
   }) async {
     state = true;
     final user = _ref.read(currentUserDetailsProvider).value!;
-    final kelompok = _ref.read(searchUserProvider(muridId)).value!.kelompok;
-
+    final kelompok = _ref.read(searchUserProvider(muridId)).value?.kelompok;
+    if (kelompok == null) {
+      showSnackBar(context, 'Tekan lagi');
+      state = false;
+      return;
+    }
     if (image != null) {
       // Upload new image if provided
       await _hkAPI.deleteImage(imageId);
